docker/requirements.txt
# ===================================================================
# REQUIREMENTS.TXT - Backend Hidrológico Censipam Porto Velho
# Baseado na análise completa do código fonte do projeto
# ===================================================================

# === BIBLIOTECAS PRINCIPAIS ===
# Data processing e análise
pandas==2.2.2
numpy==1.26.4

# Web automation (essencial para download do Hidroweb)
playwright==1.45.0

# Database connectivity (PostgreSQL corporativo)
psycopg2-binary==2.9.10

# === INTERFACE GRÁFICA (para compatibilidade, mesmo sendo local) ===
# CustomTkinter para interface moderna
customtkinter==5.2.2

# Pillow para processamento de imagens/ícones
Pillow==11.0.0

# === BIBLIOTECAS BUILT-IN (já incluídas no Python, mas listadas para referência) ===
# Estas NÃO precisam ser instaladas via pip, mas estão no seu código:
# - tkinter (built-in)
# - threading (built-in) 
# - json (built-in)
# - os (built-in)
# - sys (built-in)
# - pathlib (built-in)
# - re (built-in)
# - datetime (built-in)
# - logging (built-in)
# - shutil (built-in)
# - getpass (built-in)
# - glob (built-in)

# === UTILITÁRIAS E SUPORTE ===
# Manipulação de datas
python-dateutil==2.8.2

# Processamento de arquivos Excel (se necessário)
openpyxl==3.1.2
xlrd==2.0.1

# Validação e configuração JSON
jsonschema==4.23.0

# ===================================================================
# OBSERVAÇÕES IMPORTANTES:
# 1. tkinter: Built-in no Python, não precisa instalar
# 2. customtkinter: Usado na interface LOCAL, não no container
# 3. Pillow: Usado para ícones da interface, mantido para compatibilidade
# 4. psycopg2-binary: Versão binária para facilitar instalação Docker
# 5. playwright: Inclui automaticamente os navegadores necessários
# ===================================================================